more post revision issues
git-svn-id: file:///Users/michaelkoepke/svn/semiologic/plugins/redirect-manager/trunk@159 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ redirect-manager-admin.php
@@ -74,6 +74,10 @@ class redirect_manager_admin
 
 	function save_entry($post_ID)
 	{
+		$post = get_post($post_ID);
+		
+		if ( $post->post_type == 'revision' ) return;
+		
 		if ( current_user_can('edit_pages') || current_user_can('publish_posts') )
 		{
 			delete_post_meta($post_ID, '_redirect_url');
